Data-Visualization-with-PandasAI/requirements.txt
pandasai
streamlit
langchain_anthropic
seaborn
python-dotenv 